small fixings and explicit implements
